接入各玩法公共接线：封包、操作码、角色存档与库存路由。

+++ gms-server/src/main/java/org/gms/client/processor/action/SpawnPetProcessor.java
@@ -40,7 +40,7 @@ import java.awt.*;
 public class SpawnPetProcessor {
     private static final DataProvider dataRoot = DataProviderFactory.getDataProvider(WZFiles.ITEM);
 
-    public static void processSpawnPet(Client c, byte slot, boolean lead) {
+    public static void processSpawnPet(Client c, short slot, boolean lead) {
         if (c.tryacquireClient()) {
             try {
                 Character chr = c.getPlayer();
